moves saml header file logic to ssoe_settings_service
